Completed reintegration of old PPH tables, seperated PPH Salesforce tables
