+++ server/package.json
@@ -22,7 +22,7 @@
     "npm": ">=8.0.0"
   },
   "dependencies": {
-    "@prisma/client": "^5.1.1",
+    "@prisma/client": "^5.5.0",
     "@types/supertest": "^2.0.12",
     "class-validator": "^0.14.0",
     "cookie-parser": "^1.4.6",
@@ -61,7 +61,7 @@
     "eslint": "^8.46.0",
     "eslint-config-google": "^0.14.0",
     "nodemon": "^3.0.1",
-    "prisma": "^5.1.1",
+    "prisma": "^5.5.0",
     "supertest": "^6.3.3",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
