+++ server/prisma/seed.ts
@@ -1,5 +1,4 @@
 const {PrismaClient} = require('@prisma/client');
-const importAccounts = require('./src/seedAccounts');
 const importContacts = require('./src/seedContacts');
 const importDates = require('./src/seedDates');
 const importDiscounts = require('./src/seedDiscounts');
@@ -7,20 +6,24 @@ const importDonations = require('./src/seedDonations');
 const importEvents = require('./src/seedEvents');
 const importEventInstances = require('./src/seedEventInstances');
 const importEventTickets = require('./src/seedEventTickets');
-const importNotes = require('./src/seedNotes');
-const importOpportunity = require('./src/seedOpportunity');
 const importOrders = require('./src/seedOrders');
-const importRecordTypes = require('./src/seedRecordTypes');
 const importSeasons = require('./src/seedSeasons');
 const importSeasonTickets = require('./src/seedSeasonTickets');
-const importTicketOrders = require('./src/seedTicketOrders');
-const importTicketOrderItems = require('./src/seedTicketOrderItems');
 const importSeasonTicketTypes = require('./src/seedSeasonTicketTypes');
 const importTicketRestrictions = require('./src/seedTicketRestrictions');
 const importTicketTypes = require('./src/seedTicketTypes');
-const importTransactions = require('./src/seedTransactions');
 const importUsers = require('./src/seedUsers');
 const importOrderItems = require('./src/seedOrderItems');
+// PPH IMPORT HERE
+const importPPHAccounts = require('./src/seedPPH_Accounts');
+const importPPHContacts = require('./src/seedPPH_Contacts');
+const importPPHEvents = require('./src/seedPPH_Events');
+const importPPHNotes = require('./src/seedPPH_Notes');
+const importPPHOpportunity = require('./src/seedPPH_Opportunity');
+const importPPHRecordTypes = require('./src/seedPPH_RecordTypes');
+const importPPHTicketOrders = require('./src/seedPPH_TicketOrders');
+const importPPHTicketOrderItems = require('./src/seedPPH_TicketOrderItems');
+const importPPHTransactions = require('./src/seedPPH_Transactions');
 
 
 const prisma = new PrismaClient();
@@ -29,28 +32,31 @@ const prisma = new PrismaClient();
  * Seed database
  */
 async function main() {
-  // await importDates(prisma);
-  // await importUsers(prisma);
-  await importAccounts(prisma);
+  await importDates(prisma);
+  await importUsers(prisma);
   await importContacts(prisma);
-  // await importDonations(prisma);
-  // await importSeasons(prisma);
-  // await importDiscounts(prisma);
-  // await importTicketTypes(prisma);
+  await importDonations(prisma);
+  await importSeasons(prisma);
+  await importDiscounts(prisma);
+  await importTicketTypes(prisma);
   await importEvents(prisma);
-  await importNotes(prisma);
-  // await importEventInstances(prisma);
-  // await importTicketRestrictions(prisma);
-  // await importEventTickets(prisma);
-  // await importOrders(prisma);
-  // await importOrderItems(prisma);
-  await importRecordTypes(prisma);
-  // await importSeasonTicketTypes(prisma);
-  // await importSeasonTickets(prisma);
-  await importOpportunity(prisma);
-  await importTicketOrders(prisma);
-  await importTicketOrderItems(prisma);
-  await importTransactions(prisma);
+  await importEventInstances(prisma);
+  await importTicketRestrictions(prisma);
+  await importEventTickets(prisma);
+  await importOrders(prisma);
+  await importOrderItems(prisma);
+  await importSeasonTicketTypes(prisma);
+  await importSeasonTickets(prisma);
+  // PPH SEEDING HERE
+  await importPPHAccounts(prisma);
+  await importPPHContacts(prisma);
+  await importPPHEvents(prisma);
+  await importPPHNotes(prisma);
+  await importPPHRecordTypes(prisma);
+  await importPPHOpportunity(prisma);
+  await importPPHTicketOrders(prisma);
+  await importPPHTicketOrderItems(prisma);
+  await importPPHTransactions(prisma);
 }
 
 main()
