+++ server/prisma/src/seedEvents.ts
@@ -14,25 +14,17 @@ async function seedEvents(prisma: PrismaClient) {
       return;
     }
 
-    const yamlData = fs.readFileSync('./prisma/yaml-seeder-data/events.yaml', 'utf8');
+    const yamlData = fs.readFileSync('./prisma/legacy-data/events.yaml', 'utf8');
     const data: any[] = yaml.load(yamlData);
 
     const preparedData = data.map((item) => ({
-      event_id: item.event_id,
-      owner_id: item.owner_id,
-      name: item.name,
-      create_date: item.create_date ? new Date(item.create_date) : new Date(),
-      last_modified_date: item.last_modified_date ? new Date(item.last_modified_date) : new Date(),
-      active_flag: item.active_flag || 0,
-      description: item.description,
-      detail: item.detail,
-      event_category: item.event_category,
-      season: item.season,
-      performance_date: item.performance_date ? new Date(item.performance_date) : new Date(),
-      pre_post_show_email_flag: isNaN(item.pre_post_show_email_flag) ? null : item.pre_post_show_email_flag,
-      pre_show_email_cutoff_minutes: isNaN(item.pre_show_email_cutoff_minutes) ? null : item.pre_show_email_cutoff_minutes || 0,
-      pre_show_email_minutes: isNaN(item.pre_show_email_minutes) ? null : item.pre_show_email_minutes || 0,
-      run_time: item.run_time,
+      eventid: item.eventid,
+      seasonid_fk: item.seasonid_fk,
+      eventname: item.eventname,
+      eventdescription: item.eventdescription,
+      active: item.active,
+      seasonticketeligible: item.seasonticketeligible,
+      imageurl: item.imageurl,
     }));
 
     await prisma.events.createMany({
@@ -41,7 +33,7 @@ async function seedEvents(prisma: PrismaClient) {
 
     console.log('Events seeding completed.');
   } catch (error) {
-    console.error('Failed to seed events:', error);
+    console.error(error);
   }
 }
